start putting together the project
For now, this is a matter of just parsing and printing all the keys
of the data object. These, together, form a comprehensive list of the
names of the prototypes in the game.

Later we can work on transforming Lua data types into Rust data types,
and then follow up by constructing real data models for everything.