finish handling parsing
ultimately, the problem was that we'd unintentionally converted a bunch
of lua integers into floating-point values in the JSON, which then
caused problems when I wanted to parse them as integers in the data.

`jq` wasn't helping, because it silently converts `N.0` inputs to `N`
outputs for all numeric `N`.

With that sorted, it looks deserialization is working properly for
all vanilla recipes now. Roundtrip tests fail more often than not,
but it looks like this is in every case becasue we always serialize
recipe durations as floats, never as integers.